+++ src/Parallel.php
@@ -9,13 +9,20 @@ use Parallel\Output\Output;
 use Parallel\Output\TableOutput;
 use Parallel\TaskStack\StackedTask;
 use Parallel\TaskStack\TaskStack;
+use Parallel\TaskStack\TaskStackFactory;
+use Psr\Log\LoggerAwareTrait;
+use Psr\Log\NullLogger;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Process\Process;
+use Parallel\Exception\TaskStackFactoryException;
+use Exception;
 
 class Parallel
 {
+    use LoggerAwareTrait;
+
     /** @var string */
     private $binDirPath;
 
@@ -28,6 +35,13 @@ class Parallel
     /** @var string */
     private $logDir;
 
+    /** @var array */
+    private $tasksData = [];
+
+    /** @var TaskStackFactory */
+    private $taskStackFactory;
+
+    /** @var TaskStack */
     private $taskStack;
 
     /** @var Output */
@@ -39,22 +53,28 @@ class Parallel
     /** @var array */
     private $data = [];
 
+    /** @var string|null */
+    private $logFile;
+
     /**
-     * @param string $binDirPath
-     * @param string $fileName
-     * @param int $concurrent
+     * @param string $binDirPath    Path to directory with parallel binary
+     * @param string $fileName      Parallel binary filename
+     * @param int $concurrent       Max count of concurrent tasks
+     * @param string $logDir
      */
-    public function __construct(string $binDirPath, string $fileName, int $concurrent = 3)
+    public function __construct(string $binDirPath, string $fileName, int $concurrent = 3, string $logDir = '')
     {
         $this->binDirPath = $binDirPath;
         $this->fileName = $fileName;
         $this->concurrent = $concurrent;
+        $this->logDir = rtrim($logDir, '/');
+        $this->logger = new NullLogger();
         $this->app = new Application();
         $this->app->add(new RunCommand($this));
         $this->app->add(new AnalyzeGraphCommand($this));
 
         // Defaults
-        $this->taskStack = new TaskStack();
+        $this->taskStackFactory = new TaskStackFactory();
         $this->output = new TableOutput();
     }
 
@@ -70,9 +90,11 @@ class Parallel
 
     /**
      * @return TaskStack
+     * @throws TaskStackFactoryException
      */
     public function getTaskStack(): TaskStack
     {
+        $this->initializeTaskStack();
         return $this->taskStack;
     }
 
@@ -87,33 +109,43 @@ class Parallel
     }
 
     /**
-     * @param string $logDir
-     * @return Parallel
-     */
-    public function setLogDir(string $logDir): Parallel
-    {
-        $this->logDir = $logDir;
-        return $this;
-    }
-
-    /**
      * @param Task $task
      * @param array $runAfter
+     * @param int|null $maxConcurrentTasksCount
      * @return Parallel
      */
-    public function addTask(Task $task, $runAfter = []): Parallel
+    public function addTask(Task $task, $runAfter = [], ?int $maxConcurrentTasksCount = null): Parallel
     {
-        $this->taskStack->addTask($task, $runAfter);
+        $task->setLogger($this->logger);
+        $this->tasksData[] = [
+            'task' => $task,
+            'runAfter' => $runAfter,
+            'maxConcurrentTasksCount' => $maxConcurrentTasksCount
+        ];
         $this->app->add($task);
         return $this;
     }
 
     /**
      * Start parallel
+     * @throws Exception
      */
     public function runConsoleApp(): void
     {
         $this->app->run();
+    }
+
+    /**
+     * @param array $subnets
+     * @throws TaskStackFactoryException
+     */
+    private function initializeTaskStack(array $subnets = []): void
+    {
+        if ($this->taskStack !== null) {
+            return;
+        }
+
+        $this->taskStack = $this->taskStackFactory->create($this->tasksData, $subnets);
     }
 
     /**
@@ -122,8 +154,18 @@ class Parallel
      */
     public function execute(InputInterface $input, OutputInterface $output): void
     {
+        try {
+            $this->initializeTaskStack($input->getOption('subnet'));
+        } catch (TaskStackFactoryException $e) {
+            $this->output->errorMessage($output, $e->getMessage());
+            return;
+        }
+
         $start = microtime(true);
         $this->output->startMessage($output);
+
+        @mkdir(sprintf("%s/stats", $this->logDir));
+        $this->logFile = sprintf("%s/stats/%s.log", $this->logDir, time());
 
         // Add all tasks to tasks data array
         foreach ($this->taskStack->getStackedTasks() as $stashedTask) {
@@ -135,12 +177,24 @@ class Parallel
         while (!$this->taskStack->isEmpty()) {
             foreach ($processes as $runningProcessKey => $runningProcess) {
                 if (!$runningProcess['process']->isRunning()) {
-                    $this->taskStack->markDone($runningProcess['stackedTask']->getTask()->getName());
-                    $this->moveTaskDataToBottom($runningProcess['stackedTask']);
+                    /** @var StackedTask $doneStackedTask */
+                    $doneStackedTask = $runningProcess['stackedTask'];
+
+                    $this->taskStack->markDone($doneStackedTask->getTask()->getName());
+                    $this->moveTaskDataToBottom($doneStackedTask);
                     unset($processes[$runningProcessKey]);
+
+                    foreach ($processes as $process) {
+                        /** @var StackedTask $runningStackedTask */
+                        $runningStackedTask = $process['stackedTask'];
+
+                        $doneStackedTask->runningWithStop($runningStackedTask);
+                        $runningStackedTask->runningWithStop($doneStackedTask);
+                    }
 
                     // Redraw output when task finished
                     $this->output->printToOutput($output, $this->data, microtime(true) - $start);
+                    $this->printTaskStatsToFile($doneStackedTask);
                 }
             }
 
@@ -149,14 +203,16 @@ class Parallel
                 continue;
             }
 
-            foreach ($this->taskStack->getRunnableTasks($this->concurrent - count($processes)) as $stackedTask) {
-                $arguments = null;
-                if ($this->logDir) {
-                    $arguments = '--log_dir=' . $this->logDir;
+            foreach ($this->taskStack->getRunnableTasks($this->concurrent - count($processes), count($processes)) as $stackedTask) {
+                foreach ($processes as $process) {
+                    /** @var StackedTask $runningStackedTask */
+                    $runningStackedTask = $process['stackedTask'];
+                    $stackedTask->runningWithStart($runningStackedTask);
+                    $runningStackedTask->runningWithStart($stackedTask);
                 }
 
                 $processes[] = $process = [
-                    'process' => new Process('php ' . $this->fileName . ' ' . $stackedTask->getTask()->getName() . ' ' . $arguments, $this->binDirPath, null, null, null),
+                    'process' => new Process('php ' . $this->fileName . ' ' . $stackedTask->getTask()->getName(), $this->binDirPath, null, null, null),
                     'stackedTask' => $stackedTask
                 ];
 
@@ -170,7 +226,7 @@ class Parallel
                         ]);
 
                         foreach ($lines as $errorLine) {
-                            $this->logToFile($stackedTask->getTask()->getSanitizedName() . ': ' . StringHelper::sanitize($errorLine), 'error');
+                            $this->logger->error($stackedTask->getTask()->getSanitizedName() . ': ' . StringHelper::sanitize($errorLine));
                         }
 
                         $this->output->printToOutput($output, $this->data, microtime(true) - $start);
@@ -184,7 +240,7 @@ class Parallel
                     foreach (explode(';', $lastLine) as $statement) {
                         $explodedStatement = explode(':', $statement, 2);
                         if (count($explodedStatement) != 2) {
-                            $this->logToFile($stackedTask->getTask()->getSanitizedName() . ': Cannot parse statement: ' . $statement, 'error');
+                            $this->logger->error($stackedTask->getTask()->getSanitizedName() . ': Cannot parse statement: ' . $statement);
                             $this->buildTaskData($stackedTask, [
                                 'code_errors_count' => 1
                             ]);
@@ -235,15 +291,42 @@ class Parallel
     }
 
     /**
-     * @param string $line
-     * @param string $type
+     * @param StackedTask $stackedTask
      */
-    protected function logToFile(string $line, string $type): void
+    private function printTaskStatsToFile(StackedTask $stackedTask): void
     {
-        if (!$this->logDir) {
+        if (empty($this->logDir)) {
             return;
         }
 
-        file_put_contents($this->logDir . '/parallel_' . $type, "\n[" . date('d.m.Y H:i:s') . ']: ' . $line, FILE_APPEND);
+        $fileContent = @file_get_contents($this->logFile);
+        $data = $fileContent ? json_decode($fileContent, true) : [];
+
+        $withTasks = [];
+        if (count($stackedTask->getRunningWith())) {
+            foreach ($stackedTask->getRunningWith() as $name => $value) {
+                $withTasks[$name] = [
+                    'from' => $value['from']->format('c'),
+                    'to' => $value['to']->format('c')
+                ];
+            }
+        }
+
+        /** @var TaskData $taskData */
+        $taskData = $this->data[$stackedTask->getTask()->getName()];
+
+        $data['tasks_count'] = count($this->data);
+        $data['finished'] = $this->taskStack->isEmpty();
+        $data['tasks'][$stackedTask->getTask()->getName()] = [
+            'start_at' => $stackedTask->getStartAt()->format('c'),
+            'end_at' =>  $stackedTask->getFinishedAt()->format('c'),
+            'duration' => $taskData->getDuration(),
+            'count' => $taskData->getCount(),
+            'memory_peak' => $taskData->getMemoryPeak(),
+            'extra' => $taskData->getAllExtra(),
+            'with_tasks' => $withTasks
+        ];
+
+        file_put_contents($this->logFile, json_encode($data));
     }
 }