+++ src/TaskResult/ErrorResult.php
@@ -2,8 +2,23 @@
 
 namespace Parallel\TaskResult;
 
+use Throwable;
+
 class ErrorResult extends BaseTaskResult
 {
+    /** @var Throwable|null */
+    private $throwable;
+
+    /**
+     * @param string $message
+     * @param Throwable|null $throwable
+     */
+    public function __construct(string $message = '', Throwable $throwable = null)
+    {
+        parent::__construct($message);
+        $this->throwable = $throwable;
+    }
+
     /**
      * Result return code
      * @return int
@@ -11,5 +26,21 @@ class ErrorResult extends BaseTaskResult
     public function getCode(): int
     {
         return 100;
+    }
+
+    /**
+     * @return string
+     */
+    public function getShortName(): string
+    {
+        return 'error';
+    }
+
+    /**
+     * @return null|Throwable
+     */
+    public function getThrowable(): ?Throwable
+    {
+        return $this->throwable;
     }
 }