+++ src/TaskStack/TaskStack.php
@@ -28,11 +28,12 @@ class TaskStack
     /**
      * @param Task $task
      * @param array $runAfter
+     * @param int|null $maxConcurrentTasksCount
      * @return TaskStack
      */
-    public function addTask(Task $task, $runAfter = []): TaskStack
+    public function addTask(Task $task, $runAfter = [], ?int $maxConcurrentTasksCount = null): TaskStack
     {
-        $this->stackedTasks[$task->getName()] = new StackedTask($task, $runAfter);
+        $this->stackedTasks[$task->getName()] = new StackedTask($task, $runAfter, $maxConcurrentTasksCount);
         $this->tasksCount++;
         return $this;
     }
@@ -71,13 +72,36 @@ class TaskStack
     /**
      * Get desired number of runnable tasks
      * @param int $count
+     * @param int $currentTasksRunningCount
      * @return StackedTask[]
      */
-    public function getRunnableTasks(int $count = 1): array
+    public function getRunnableTasks(int $count, int $currentTasksRunningCount): array
     {
+        $runnableTasks = [];
+        $selected = 0;
+
         /** @var StackedTask[] $runnableTasks */
-        $runnableTasks = array_slice($this->runnableTasks, 0, $count);
-        $this->runnableTasks = array_slice($this->runnableTasks, $count);
+        foreach ($this->runnableTasks as $key => $task) {
+            if ($selected === $count) {
+                break;
+            }
+
+            // Check if any running task reach max concurrent task count
+            foreach ($this->runningTasks as $runningTask) {
+                if ($runningTask->getMaxConcurrentTasksCount() !== null && ($runningTask->getMaxConcurrentTasksCount() < $currentTasksRunningCount + count($runnableTasks))) {
+                    break 2;
+                }
+            }
+
+            // Check if selected runnable task reach max concurrent task count
+            if ($task->getMaxConcurrentTasksCount() !== null && ($task->getMaxConcurrentTasksCount() < $currentTasksRunningCount + count($runnableTasks))) {
+                continue;
+            }
+
+            $runnableTasks[] = $task;
+            unset($this->runnableTasks[$key]);
+            $selected++;
+        }
 
         foreach ($runnableTasks as $runnableTask) {
             $this->runningTasks[$runnableTask->getTask()->getName()] = $runnableTask;